requirements.txt
playwright>=1.40.0
openai>=1.10.0
beautifulsoup4>=4.12.0
langdetect>=1.0.9
pyyaml>=6.0
pandas>=2.0.0
tqdm>=4.66.0
aiohttp>=3.9.0
lxml>=4.9.0
python-dotenv>=1.0.0
Pillow>=10.0.0
